+++ src/game/explosion.h
@@ -11,7 +11,7 @@ struct explosion
     uint8_t x;
     uint8_t y;
 
-    int radius;
+    int8_t radius;
     TCOD_ColorRGB color;
     float lifetime;
 
@@ -22,7 +22,7 @@ struct explosion *explosion_new(
     uint8_t floor,
     uint8_t x,
     uint8_t y,
-    int radius,
+    int8_t radius,
     TCOD_ColorRGB color,
     struct actor *initiator);
 void explosion_delete(struct explosion *explosion);
